Implement v1 image tarball generation in new legacy package
