+++ pkg/v1/config.go
@@ -35,8 +35,6 @@ type ConfigFile struct {
 	Config          Config    `json:"config"`
 	ContainerConfig Config    `json:"container_config,omitempty"`
 	OSVersion       string    `json:"osversion,omitempty"`
-	ID              string    `json:"id,omitempty"`
-	Parent          string    `json:"parent,omitempty"`
 	Throwaway       bool      `json:"throwaway,omitempty"`
 }
